# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "{{.Description}}",
        "title": "{{.Title}}",
        "contact": {},
        "license": {},
        "version": "{{.Version}}"
    },
    "host": "{{.Host}}",
    "basePath": "{{.BasePath}}",
    "paths": {
...
        "/users": {
            "get": {
                "description": "获取用户列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "获取用户列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "With the bearer started JWT",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "起始数",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "获取数",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "用户状态 1启用 2禁用 3已过期",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "用户到期起始时间",
                        "name": "starttime",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "用户到期终止时间",
                        "name": "endtime",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "用户名",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "群组id",
                        "name": "ownid",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "用户id",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "用户级别",
                        "name": "permisson",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "VIP等级",
                        "name": "vip",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"Error\":\"Success\",\"Data\": object}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "301": {
                        "description": "{\"Error\":\"Re-login\",\"Data\": object}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "{\"Error\":\"error\",\"Data\": null}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": ".添加用户",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": ".添加用户(仅管理员以上)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "With the bearer started JWT",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "新用户信息",
                        "name": "NewUserinfo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/api.NewUserinfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"Error\":\"Success\",\"Data\": object}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "301": {
                        "description": "{\"Error\":\"Re-login\",\"Data\": object}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "{\"Error\":\"error\",\"Data\": null}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": ".删除用户",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": ".删除用户(仅管理员以上)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "With the bearer started JWT",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "用户ID",
                        "name": "userDelID",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/api.UserDelID"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"Error\":\"Success\",\"Data\": object}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "301": {
                        "description": "{\"Error\":\"Re-login\",\"Data\": object}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "{\"Error\":\"error\",\"Data\": null}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/users/expire": {
            "put": {
                "description": ".更改用户过期时间",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": ".更改用户过期时间(仅管理员以上)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "With the bearer started JWT",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "用户过期时间信息",
                        "name": "userExpire",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/api.UserExpire"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"Error\":\"Success\",\"Data\": object}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "301": {
                        "description": "{\"Error\":\"Re-login\",\"Data\": object}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "{\"Error\":\"error\",\"Data\": null}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/users/info": {
            "put": {
                "description": ".更改用户信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": ".更改用户信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "With the bearer started JWT",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "用户信息",
                        "name": "Info",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/api.Info"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"Error\":\"Success\",\"Data\": object}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "301": {
                        "description": "{\"Error\":\"Re-login\",\"Data\": object}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "{\"Error\":\"error\",\"Data\": null}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/users/password": {
            "put": {
                "description": ".更改用户密码",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": ".更改用户密码",
                "parameters": [
                    {
                        "type": "string",
                        "description": "With the bearer started JWT",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "用户密码",
                        "name": "PasswordInfo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/api.PasswordInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"Error\":\"Success\",\"Data\": object}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "301": {
                        "description": "{\"Error\":\"Re-login\",\"Data\": object}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "{\"Error\":\"error\",\"Data\": null}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/users/status": {
            "put": {
                "description": "更改用户状态",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "更改用户状态(仅管理员以上)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "With the bearer started JWT",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "用户状态信息",
                        "name": "userStatus",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/api.UserStatus"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"Error\":\"Success\",\"Data\": object}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "301": {
                        "description": "{\"Error\":\"Re-login\",\"Data\": object}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "{\"Error\":\"error\",\"Data\": null}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.Info": {
            "type": "object",
            "properties": {
                "address": {
                    "description": "地址 不修改不传",
                    "type": "string"
                },
                "email": {
                    "description": "邮箱 不修改不传",
                    "type": "string"
                },
                "id": {
                    "description": "用户ID 修改自身不传",
                    "type": "integer"
                },
                "mobile": {
                    "description": "手机 不修改不传",
                    "type": "string"
                },
                "photo": {
                    "description": "头像  不修改不传",
                    "type": "string"
                }
            }
        },
...
        "api.NewUserinfo": {
            "type": "object",
            "properties": {
                "details": {
                    "description": "备注",
                    "type": "string"
                },
                "expired": {
                    "description": "过期时间",
                    "type": "integer"
                },
                "name": {
                    "description": "用户名",
                    "type": "string"
                },
                "password": {
                    "description": "用户密码",
                    "type": "string"
                },
                "status": {
                    "description": "状态 1启用 2禁用",
                    "type": "integer"
                },
                "vip": {
                    "description": "Vip 等级",
                    "type": "integer"
                }
            }
        },
        "api.PasswordInfo": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "用户ID 修改自身不传",
                    "type": "integer"
                },
                "oldpassword": {
                    "description": "旧密码 MD5加密 修改下级用户不传",
                    "type": "string"
                },
                "password": {
                    "description": "新密码 MD5加密",
                    "type": "string"
                }
            }
        },
...
        "api.UserDelID": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "用户ID",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "api.UserExpire": {
            "type": "object",
            "properties": {
                "expiredtime": {
                    "description": "过期时间",
                    "type": "integer"
                },
                "ids": {
                    "description": "用户ID",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "api.UserStatus": {
            "type": "object",
            "properties": {
                "ids": {
                    "description": "用户ID",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "status": {
                    "description": "状态",
                    "type": "integer"
                }
            }
        }
    }
}